Rename model to base class for events and transaction
